+++ CHANGELOG.md
@@ -46,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix(securityhub): by default only send HIGH and CRITICAL SecurityHub events to CloudWatch. If other events are needed in CloudWatch update configuration
 - fix(pipeline): update iam policy for github token change lambda to allow pipeline update
 - fix(sample-config): extend preventative measures to roles
-
+- fix(asea): lookup tgw attachments using asea vpc names
 
 ### Changed