+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.4] - 2025-07-21
+
+### Fixed
+
+- fix(asea): lookup tgw attachments using asea vpc names
+- fix(logging): add permissions to cloudwatch log creation for kms association
+- fix(networking): remove duplicates for gateway loadbalancer endpoints
+
 ## [1.12.3] - 2025-06-11
 
 ### Added